Add styling for sign up page
