+++ client/src/assets/stylesheets/SignupForm.css
@@ -1,6 +1,6 @@
 @import 'variables.css';
 
-.login-form-container {
+.signup-form-container {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -8,7 +8,7 @@
   height: 100%;
 }
 
-.login-form-header { 
+.signup-form-header { 
   position: relative;
   display: flex;
   flex-direction: row;
@@ -19,7 +19,7 @@
   background: var(--red);
 }
 
-.login-form-header a .logo-wrapper {
+.signup-form-header a .logo-wrapper {
   position: absolute;
   top: 18px;
   left: 46.5%;
@@ -30,12 +30,12 @@
   width: 100%;
 }
 
-.login-form-header a .logo-wrapper svg {
+.signup-form-header a .logo-wrapper svg {
   width: 85px;
   height: 30px;
 }
 
-.login-form-header a .logo-wrapper svg text {
+.signup-form-header a .logo-wrapper svg text {
   font-family: 'helveticaRounded';
   font-size: 24px;
   font-weight: bold;
@@ -49,12 +49,12 @@
   paint-order: stroke fill;
 }
 
-.login-form-header a .logo-wrapper img {
+.signup-form-header a .logo-wrapper img {
   width: 25px;
   margin: -15px 0 0 -2px;
 }
 
-.login-form-body-container {
+.signup-form-body-container {
   display: flex;
   flex-direction: row;
   justify-content: center;
@@ -64,7 +64,7 @@
   padding: 30px 100px;
 }
 
-.login-main-form-container {
+.signup-main-form-container {
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -74,42 +74,56 @@
   margin-right: 100px;
 }
 
-.login-main-form-container h1 {
+.signup-main-form-container h1 {
   font-family: 'helveticaBold';
   font-size: 22px;
   color: var(--red);
 }
 
-.login-main-form-container p {
+.signup-main-form-container p {
   font-family: 'helveticaBold';
   font-size: 14px;
   color: var(--black);
 }
 
-.login-main-form-container p a {
+.signup-main-form-container p a {
   text-decoration: none;
   color: var(--blue);
 }
 
-.login-main-form-container p a:hover {
+.signup-main-form-container p a:hover {
   text-decoration: underline;
 }
 
-.login-main-form-container small {
+.signup-main-form-container small {
   font-family: 'helveticaNormal';
   font-size: 12px;
   color: var(--mid-text);
   padding: 10px 0 30px 0;
 }
 
-.login-main-form-container form {
+.signup-main-form-container form {
   display: flex;
   flex-direction: column;
   align-items: flex-end;
   width: 100%;
 }
 
-.login-main-form-container form input {
+.signup-main-form-container form .first-and-last-name-container {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 100%;
+  margin-bottom: -25px;
+}
+
+.signup-main-form-container form .first-and-last-name-container input:first-of-type,
+.signup-main-form-container form .first-and-last-name-container input:last-of-type {
+  width: 49%;
+  height: 100%;
+}
+
+.signup-main-form-container form input {
   font-family: 'helveticaNormal';
   font-size: 14px;
   color: var(--black);
@@ -121,11 +135,11 @@
   margin-bottom: 10px;
 }
 
-.login-main-form-container form input:last-of-type {
+.signup-main-form-container form input:last-of-type {
   margin-bottom: 40px;
 }
 
-.login-main-form-container form button {
+.signup-main-form-container form button {
   font-family: 'helveticaBold';
   font-size: 16px;
   background: var(--red);
@@ -135,29 +149,29 @@
   width: 100%;
 }
 
-.login-main-form-container form button:hover {
+.signup-main-form-container form button:hover {
   background: var(--button-hover-red);
   transition: background 0.5s;
   cursor: pointer;
 }
 
-.login-main-form-container form small {
+.signup-main-form-container form small {
   font-family: 'helveticaNormal';
   font-size: 12px;
   color: var(--subtle-text);
 }
 
-.login-main-form-container form small a {
+.signup-main-form-container form small a {
   color: var(--blue);
   text-decoration: none;
 }
 
-.login-main-form-container form small a:hover {
+.signup-main-form-container form small a:hover {
   text-decoration: underline;
   cursor: pointer;
 }
 
-.login-form-image-container {
+.form-image-container {
   display: flex;
   flex-direction: row;
   justify-content: center;
@@ -165,6 +179,18 @@
   width: 30%;
 }
 
-.login-form-image-container img {
+.form-image-container img {
+  width: 100%;
+}
+
+.signup-form-image-container {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  width: 30%;
+}
+
+.signup-form-image-container img {
   width: 100%;
 }
